# Improve String#casecmp? and Symbol#casecmp? performance with ASCII st…